Seperated svg logic in smaller methods to make it more readable
