+++ InsightsFrontend/src/app/app.component.html
@@ -1,1 +1,1 @@
-<app-release-coordinate-plane></app-release-coordinate-plane>
+<app-release-graph></app-release-graph>
