+++ InsightsFrontend/src/app/pages/release-graph/release-graph.component.html
@@ -1,5 +1,6 @@
 <div class="graph-container">
-	<svg #svgElement
+	<svg id="svgElement"
+		 #svgElement
 		 (wheel)="onWheel($event)"
 		 (mousedown)="onMouseDown($event)"
 		 (mousemove)="onMouseMove($event)"
@@ -9,15 +10,17 @@
 		<g [attr.transform]="'translate(' + translateX + ',' + translateY + ') scale(' + scale + ')'">
 			@for (releaseNode of releaseNodes; track releaseNode) {
 				<g [attr.transform]="'translate(' + releaseNode.position.x + ',' + releaseNode.position.y + ')'">
-					<circle r="20" [attr.fill]="releaseNode.color" stroke="#333" stroke-width="2"></circle>
-					<text text-anchor="middle" y="-30" fill="black" font-size="14">
+					<circle r="20" [attr.fill]="releaseNode.color"></circle>
+					<text y="-30">
 						{{ releaseNode.label }}
 					</text>
 				</g>
 			}
 
 			@for (releaseLink of releaseLinks; track releaseLink) {
-				<path class="line" stroke="#999" stroke-width="2" fill="none" [attr.d]="getCustomPath(releaseLink)"></path>
+				<path
+					[attr.d]="getCustomPath(releaseLink)">
+				</path>
 			}
 		</g>
 	</svg>
