+++ InsightsFrontend/src/app/pages/release-graph/release-graph.component.scss
@@ -20,10 +20,20 @@
 
 	circle {
 		transition: transform 0.2s ease-out;
+		stroke: #EFEFEF;
+		stroke-width: 5;
 	}
 
 	text {
 		font-size: 12px;
 		font-weight: bold;
+		text-anchor: middle;
+
+	}
+
+	path {
+		stroke: #999;
+		stroke-width: 2;
+		fill: none;
 	}
 }
